+++ CHANGELOG.md
@@ -7,7 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename shared `AllocErrorType` trait to `AllocError`
+
 ## [0.10.0] - 2026-02-11
+
+### Added
+
+- `AllocErrorType` trait shared between `alloc_mut` and `alloc` traits
 
 ### Changed